added all dynamoose models to separate folder
