+++ server/models/ride.ts
@@ -1,0 +1,38 @@
+import dynamoose from 'dynamoose';
+
+export enum Type {
+  ACTIVE = 'active',
+  PAST = 'past',
+  UNSCHEDULED = 'unscheduled',
+}
+
+export type RideType = {
+  type: Type,
+  id: string,
+  startLocation: string,
+  endLocation: string,
+  startTime: string,
+  endTime: string,
+  riderId: string,
+  driverId?: string,
+};
+
+const schema = new dynamoose.Schema({
+  type: {
+    hashKey: true,
+    type: String,
+    enum: Object.values(Type),
+  },
+  id: {
+    rangeKey: true,
+    type: String,
+  },
+  startLocation: String,
+  endLocation: String,
+  startTime: String,
+  endTime: String,
+  riderId: String,
+  driverId: String,
+});
+
+export const Ride = dynamoose.model('Rides', schema, { create: false });
