+++ server/router/dispatcher.ts
@@ -1,37 +1,11 @@
 import express from 'express';
 import { v4 as uuid } from 'uuid';
-import dynamoose from 'dynamoose';
 import * as db from './common';
+import { Dispatcher } from '../models/dispatcher';
 
 const router = express.Router();
 
-// only an admin should be able to add another dispatcher
-type AccessLevel = 'Admin' | 'SDS' | 'Dispatcher'
-
-type DispatcherType = {
-  id: string,
-  firstName: string,
-  lastName: string,
-  phoneNumber: string,
-  email: string,
-  accessLevel: AccessLevel
-};
-
-const schema = new dynamoose.Schema({
-  id: String,
-  firstName: String,
-  lastName: String,
-  phoneNumber: String,
-  email: String,
-  accessLevel: {
-    type: String,
-    enum: ['Admin', 'SDS', 'Dispatcher'],
-  },
-});
-
 const tableName = 'Dispatchers';
-
-const Dispatcher = dynamoose.model(tableName, schema, { create: false });
 
 // Put a driver in Dispatchers table
 router.post('/', (req, res) => {
