+++ server/router/location.ts
@@ -1,25 +1,11 @@
 import express from 'express';
 import { v4 as uuid } from 'uuid';
-import dynamoose from 'dynamoose';
 import * as db from './common';
+import { Location } from '../models/location';
 
 const router = express.Router();
 
-export type LocationType = {
-  id: string,
-  name: string,
-  address: string,
-};
-
-const schema = new dynamoose.Schema({
-  id: String,
-  name: String,
-  address: String,
-});
-
 const tableName = 'Locations';
-
-export const Location = dynamoose.model(tableName, schema, { create: false });
 
 // Get a location by id in Locations table
 router.get('/:id', (req, res) => {
